# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/pds-backend/api/getTransferList": {
            "get": {
                "tags": [
                    "PDS User Transfer"
                ],
                "summary": "Transfer List",
                "description": "Total Transfer List",
                "operationId": "getTransferList",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/addTransferRecord": {
            "post": {
                "tags": [
                    "PDS User Transfer"
                ],
                "summary": "Add New Transfer Record",
                "description": "Add New New Transfer Record",
                "operationId": "addTransferRecord",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "to_office": {
                                        "type": "text"
                                    },
                                    "from_office": {
                                        "type": "text"
                                    },
                                    "to_department": {
                                        "type": "text"
                                    },
                                    "from_department": {
                                        "type": "text"
                                    },
                                    "to_designation": {
                                        "type": "text"
                                    },
                                    "from_designation": {
                                        "type": "text"
                                    },
                                    "transfer_order": {
                                        "type": "text"
                                    },
                                    "transfer_order_number": {
                                        "type": "text"
                                    },
                                    "transfer_type": {
                                        "type": "text"
                                    },
                                    "transfer_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "join_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "transfer_letter": {
                                        "type": "text"
                                    },
                                    "status": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Added  Transfer Record Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/updateTransferRecord/{id}": {
            "post": {
                "tags": [
                    "PDS User Transfer"
                ],
                "summary": "Update Transfer Record",
                "description": "Update Transfer Record",
                "operationId": "updateTransferRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "to_office": {
                                        "type": "text",
                                        "example": 1
                                    },
                                    "from_office": {
                                        "type": "text",
                                        "example": 1
                                    },
                                    "to_department": {
                                        "type": "text",
                                        "example": 1
                                    },
                                    "from_department": {
                                        "type": "text",
                                        "example": 1
                                    },
                                    "to_designation": {
                                        "type": "text",
                                        "example": 1
                                    },
                                    "from_designation": {
                                        "type": "text",
                                        "example": 1
                                    },
                                    "transfer_order": {
                                        "type": "text",
                                        "example": "Order123"
                                    },
                                    "transfer_order_number": {
                                        "type": "text",
                                        "example": "1204"
                                    },
                                    "transfer_type": {
                                        "type": "text",
                                        "example": 12
                                    },
                                    "transfer_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-11"
                                    },
                                    "join_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-11"
                                    },
                                    "transfer_letter": {
                                        "type": "text",
                                        "example": "letter.pdf"
                                    },
                                    "status": {
                                        "type": "text",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated Transfer Record Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/deleteTransferRecord/{id}": {
            "delete": {
                "tags": [
                    "PDS User Transfer"
                ],
                "summary": "Delete Transfer Record",
                "description": "Delete  Transfer Record With Valid ID",
                "operationId": "deleteTransferRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully, Delete Album"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/activeTransferRecord/{id}": {
            "patch": {
                "tags": [
                    "PDS User Transfer"
                ],
                "summary": "Active Transfer Record",
                "description": "Active Specific Transfer Record With Valid ID",
                "operationId": "activeTransferRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully, Active Album"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}